P1-1 economy integrity + P1-3 cookie flags
P1-1 (H5/H12) — validated against Postgres 13:
- Atomic conditional debit (UPDATE ... WHERE balance >= :amt) in EconomyService;
  DecrementCurrency fails closed on insufficient funds. Subsumes SELECT FOR UPDATE.
- Migration: clamp negatives -> widen balances int->bigint -> CHECK(balance>=0) on
  user_economy + group_economy. Verified: over-spend debits 0 rows, forced negative
  rejected by CHECK, full 88-migration chain applies clean.
- Widen UserEconomy DTO int->long (H12).
- Economy lock in CurrencyExchange.PlaceOrder (not CloseOrder — would ABBA-deadlock
  with the order lock; CloseOrder only refunds and is already order-locked).

P1-3 (H10):
- Shared SessionCookie.Append helper (HttpOnly + Secure) used by Razor login/signup
  and the v2 controller; adds the previously-missing HttpOnly to the Razor paths.
- Frontend validate-and-add-cookie.js: Secure (prod-gated) + SameSite on both cookies.

+++ services/Roblox/Roblox.Dto/Users/Economy.cs
@@ -5,8 +5,10 @@ namespace Roblox.Dto.Users;
 
 public class UserEconomy
 {
-    public int robux { get; set; }
-    public int tickets { get; set; }
+    // 64-bit to match the ledger (user_transaction.amount is long) and avoid overflow/
+    // truncation of large balances (security finding H12). Backed by bigint columns.
+    public long robux { get; set; }
+    public long tickets { get; set; }
 }
 
 public class SummaryEntryDb
